add --publish option to quasar-build
electron-builder has options to publish the new packaged executables to Github, Bintray, S3, etc. but this must be specifically triggered. Users must configure the publishing options, however just the presence of a properly-configured option set will not actually do anything unless the `electron-build` CLI command has a `--publish` flag. Our case is more complicated because we are triggering the build programmatically (https://www.electron.build/api/electron-builder), not via the CLI. So we need a `publish` key in the options that we pass to the API, that is *outside* the `config` options.

This change adds a `--publish` (or `-P`) flag that passes its value to electron-builder. It does nothing if not in electron mode or not using electron-builder.