add `publish` key to cfg.electron.builder
If the user is building for electron and using electron-builder, this is where we check whether the user specified the new `--publish` or `-P` flag. It is not read or used anywhere else, so it is totally ignored if building SPA, Cordova, etc., and is also ignored if using electron-packager. If found, its value is passed directly to the electron-builder API. If this value is not a valid option (see https://www.electron.build/configuration/publish), or if the user has not also created a valid publish configuration, then electron-builder will raise an error.